+++ src/collective/nitf/setuphandlers.py
@@ -144,3 +144,20 @@ def upgrade_to_1007(context, logger=None):
     profile = 'profile-collective.nitf:upgrade_to_1007'
     setup = getToolByName(context, 'portal_setup')
     setup.runAllImportStepsFromProfile(profile)
+
+
+def upgrade_to_1008(context, logger=None):
+    """
+    """
+    if logger is None:
+        # Called as upgrade step: define our own logger
+        logger = logging.getLogger(PROJECTNAME)
+
+    dependencies = ('plone.app.relationfield', 'plone.app.intid')
+    for dependency in dependencies:
+        qi = getToolByName(context, 'portal_quickinstaller')
+        if not qi.isProductInstalled(dependency):
+            qi.installProduct(dependency)
+            logger.info("{0} was installed.".format(dependency))
+        else:
+            logger.info("{0} already installed; nothing to do.".format(dependency))